requirements.txt
ultralytics>=8.2.50
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0
torch>=2.2.0
fastapi[standard]
python-multipart
uvicorn
opencv-python
faiss-cpu
transformers
